+++ .gitignore
@@ -1,5 +1,3 @@
-# ignoring run example files
-examples/*
 
 # ignoring built test docs
 docs/build/
